Commit before a pull from the master
